CGL: 4.0 CF: bond, death and restart
1/3 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: kernel/sys.c: Introduce ALWAYS_RESTART config option
Date: Wed, 17 Mar 2010 19:48:53 -0700

If enabled, Force the system to restart instead of halt or power
off.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Integrated-by: Yong Zhang <yong.zhang@windriver.com>
]

2/3 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: Notification of death of arbitrary processes
Date: Thu, 18 Mar 2010 01:27:00 -0700

This exposes a new feature which may be called to request
notification when an arbitrary process changes state. The
caller specifies a pid, signal number, and event mask, and
when that pid dies, or is stopped, or anything else that
would normally cause a SIGCHLD, the kernel will send the
specified signal to the caller if the event is in the event
mask originally passed down. The siginfo_t struct will
contain the same information as would be included with SIGCHLD.

This is exposed to userspace via the prctl() call with the
PR_DO_NOTIFY_TASK_STATE option.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

3/3 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: bonding: Bonding device info support
Date: Fri, 19 Mar 2010 00:29:00 -0700

This provides the physical device the packet came in from
for a bonded ethernet device. (as ancillary data)

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Integrated-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>