x86: fix the bad things in the bad patch.
JimS did an audit on this, and most of it gets thrown away,
with the remainder having #ifdefs added around suspect items.
Some of the suspect items had corresponding stub functions
for the non-mc case, and would have been OK, but having the
ifdef makes it clearly safe by a casual inspection.

The now unused stubs have been removed from the corresponding
commit that has already migrated out of the multicore-cache.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>