multicore: tag mystery ppc patch as moved.
I was hesitant to move this into master because it was
an old patch from a mailing list and not an actual upstream
commit.  But since multiple BSPs were already using it in
master anyway, MarkA has deployed it globally and will be
async looking into the possibility of replacing it with
a newer upstream version to limit risk.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>