MILS: integrate changes from the embedded MILS git repo
This list of commits represents the additional changes for MILS
that were living in the MILS specific git repo that was
embedded in the multicore_mils_dev branch of wrll-multicore.

As such they have yet to be deployed in a non-MILS build
environment, but everything within should already be within
a CONFIG_WRHV_MILS anyway.

These will also largely be squished/folded out of existence.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>