Initial cleanup of the fsl_p4080
Since the fsl_p4080 BSP missed 3.0.2 release many of its patches
had to be carried by multicore. Now that we are preparing 3.0.3 we
can work to get these included in the standard product. Many of these
already are and here we are removing these patches as they are no
longer needed here.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>