add wrhvImage target
-------------
From: Paul Gortmaker <paul.gortmaker@windriver.com>
Date: Thu, 1 Apr 2010 13:04:25 -0400
Subject: [PATCH] wrhvImage: create a replica of simpleImage that we can tweak

The default load address for a simpleImage is too low, and
it will fail to boot on the WR hypervisor with an error
message saying essentially just that.  Rather than change
the defaults for simpleImage, this creates a wrhvImage
which we are free to adjust in any way we want.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
---
 arch/powerpc/Makefile      |    3 ++-
 arch/powerpc/boot/Makefile |    9 ++++++++-
 arch/powerpc/boot/wrapper  |    5 +++++
 3 files changed, 15 insertions(+), 2 deletions(-)
--------

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>