Fix kmemtrace patch context to fix patching issue
CQID:WIND00204010

Now that we always patch kmemleak into the kernel there was a bit
of context issue with the kmemtrace feature patches. No big
surprizes as it was mostly context but one change involved
removing a kmemleak binding so the features should not be
used at the same time.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>