powerpc: Add missed unlock for vio_cmo.lock
CQID:WIND00206479

1/1 [
Author: Zumeng Chen
Email: zumeng.chen@windriver.com
Subject: powerpc: Add missed unlock for vio_cmo.lock
Date: Wed, 7 Apr 2010 14:15:45 +0800

The vio_cmo.lock is taken at the start of vio_cmo_set_dev_desired
but is not unlocked in the case of early exit. To fix it, we
add the unlock before the early return.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
]

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>