arm: 4.0 carry forward
1/4 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: arm: ARM EABI socketcall
Date: Tue, 15 Jul 2008 15:49:27 -0400

This patch make that the sys_socketcall system call is available when
EABI is used. The reason is that the user-space applications in both
glibc_std rootfs and glibc_small rootfs are still using sys_socketcall
to request kernel service.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
]

2/4 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: arm: ARM XIP u-boot updates
Date: Tue, 15 Jul 2008 15:30:36 -0400

This makes the necessary changes to the ARM build process
to generate a XIP bootable image from a uboot image.

The minor modification for vmlinux.lds.S makes uxipImage to be
booted normally by inserting uboot header into Text segment of
Linux kernel, so the size of uboot header should be added
into TEXTADDR of vmlinux.lds.S.

Signed-off-by: <bruce.ashfield@windriver.com>
Signed-off-by: <zumeng.chen@windriver.com>
Integrated-by: Liming Wang <liming.wang@windriver.com>
]

3/4 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: arm: Prevent prefetch optimization for ARM
Date: Fri, 9 Apr 2010 09:06:06 +0800

Some ARM kernels have been experiencing bugs/boot issues due
to optimization. In particular, code such as "prefetch(next)" in a
loop, causes the compiler to conclude that next can never be a
NULL pointer (because it got dereferenced), and thus optimize the
loop's test into while(1).

A more complete reference is at:

http://mirror.celinuxforum.org/gitstat/sendmail.php?mail=nico@cam.org&commit=02828845dda5ccf921ab2557c6ca17b6e7fc70e2

Because the constraint to the inline asm used in the prefetch() macro
is a memory operand, gcc assumes that the asm code does dereference the
pointer and the delete-null-pointer-checks optimization kicks in.
Inspection of generated assembly for the above example shows that
bar() is indeed called unconditionally without any test on the value
of x.

Of course in the prefetch case there is no real dereference and it
cannot be assumed that a null pointer would have been caught at that
point. This causes kernel oopses with constructs like
hlist_for_each_entry() where the list's 'next' content is prefetched
before the pointer is tested against NULL, and only when gcc feels
like applying this optimization which doesn't happen all the time
with more complex code.

It appears that the way to prevent delete-null-pointer-checks
optimization to occur in this case is to make prefetch() into a static
inline function instead of a macro. At least this is what is done on
x86_64 where a similar inline asm memory operand is used (I presume
they would have seen the same problem if it didn't work) and resulting
code for the above example confirms that.

An alternative would consist of replacing the memory operand by a
register operand containing the pointer, and use the addressing mode
explicitly in the asm template. But that would be less optimal than an
offsettable memory reference.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

4/4 [
Author: Donn Seeley
Email: donn.seeley@windriver.com
Subject: arm: Arm VFP Context
Date: Tue, 15 Jul 2008 15:26:03 -0400

This patch implements context saving for the ARM VFP hardware floating
point unit in two specific circumstances:

  +     before a fork(), so that the child inherits the parent's floating
        point state; and

  +     across a caught signal, so that the use of floating point in a
        signal handler won't destroy caller-saved floating point state
        in the the interrupted context.

Few real programs actually need this functionality, but when they do
need it, the bugs can lead to data that's corrupted in a way that is
subtle and hard to identify.

Also note: the condition field "NE" in vfp instruction is not
supported. So add a instruction "beq" to replace it.

Besides, when FPEXC.EX is set to zero, the FPINST and FPINST2
registers are not implemented. So FPINST doesn't need to be restored when
FPEXC.EX is not set.

Signed-off-by: Donn Seeley <donn.seeley@windriver.com>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Stanley.Miao <stanley.miao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>