8641D mc: fix dain bramage in Makefile changes.
The Makefile entries are using an assignment and not an
additive operator.  So changing the order and/or listing
them in two places can have an effect.

Fix it to take advantage of the overriding nature of the
assignment, which cleans up things a bunch.

The handling of excluding fpu was busted too, in that
the change was made based on CONFIG_WRHV, but yet this
is a MILS specific patch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>