Correctly deal with make that has an argument which contains an 's'
1/1 [
Author: Jason Wessel
Email: jason.wessel@windriver.com
Subject: Correctly deal with make that has an argument which contains an "s"
Date: Mon, 26 Apr 2010 15:56:14 -0500

When using remake, which is based on gnumake, if you invoke
an example build as shown below, the build will become silent
due to the top level make file incorrectly guessing that
the end user wants a silent build because an argument that
contained an "s" was used.

remake --no-extended-errors

Fix up the top level Makefile to use filter with a list of
options that mean silent with the various revisions of gnumake,
instead of findstring.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
CC: Michal Marek <mmarek@suse.cz>
CC: Andrew Morton <akpm@linux-foundation.org>
CC: linux-kbuild@vger.kernel.org
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>