compilation issue building preempt_rt+early-ftrace
CQID:WIND00210686

preempt_rt adds a new parameter "preempt_count" to trace_function()
which of course requires all callers pass this parameter. This change was
handled in most cases by the patch which makes the param. addition but since
early-ftrace is added later on it doesn't have the required change. For
non-preempt_rt kernels the original param count remains thus we handle
this via an #ifndef to handle both cases.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>