cgl: fix buglet in death notify
1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: Notification of death of arbitrary processes
Date: Thu, 18 Mar 2010 01:27:00 -0700

This exposes a new feature which may be called to request
notification when an arbitrary process changes state. The
caller specifies a pid, signal number, and event mask, and
when that pid dies, or is stopped, or anything else that
would normally cause a SIGCHLD, the kernel will send the
specified signal to the caller if the event is in the event
mask originally passed down. The siginfo_t struct will
contain the same information as would be included with SIGCHLD.

This is exposed to userspace via the prctl() call with the
PR_DO_NOTIFY_TASK_STATE option.

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>