CGL: set configs to map onto previous configs
We'd set some things off in the baseline config, but the problem
was that it wasn't properly using "off" since a "=n" is
NOT an off operation.  Now, with them properly off via the
"is not set", we need to explicitly enable them for CGL to
ensure our final configuration files remain as they were.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>