nlm_xls_atx_64_be: add hugetlb support
1/1 [
Author: Jack Tan
Email: jack.tan@windriver.com
Subject: nlm_xls_atx_64_be: add hugetlb support
Date: Fri, 11 Jun 2010 16:07:28 +0800

Introduce hugetlb support to the RMI (CPU_PHOENIX) class
of boards. This is done via the introduction of <foo>_pte
and <blah>_pte that manage the page tables in a manner
specific to these boards.

This patch is based on SDK 1.7.0 and the mainline kernel.

Signed-off-by: Jack Tan <jack.tan@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>