8572: refresh
1/4 [
Author: Benjamin Walsh
Email: benjamin.walsh@windriver.com
Subject: PPC32/8572ds/kexec: add KEXEC_SMP_BOOTABLE option
Date: Fri, 21 May 2010 15:26:06 -0400

The 8572ds code currently supports two ways of waking up a secondary
CPU. The older uboot is not aware of multiple CPUs and thus leaves
the CPU off and then the kernel starts it; the newer uboot knows
about them and wakes them up and spins them on a tight loop, leaving
the kernel to make them enter the kernel by writing a start address
in the spin loop variable. The kexec-bootable kernel takes an
approach that is similar to the latter, with one more step.

When the secondary CPU goes down in the original kernel, it first
spins in a tight loop in the control code page, which is randomly
allocated, until the copy of the new kernel is done. At that point,
the primary CPU jumps to the new kernel. It then has to bring the
secondary CPU into the second kernel as well since the control code
page has a high chance of being reused by the new kernel before the
primary CPU hits the secondary wakeup code. To achieve this, the
primary CPU writes the address of a new loop in the second kernel
at which the secondary jumps to. When the wakeup code is hit by the
primary, it finally instructs the secondary to run __early_start.

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>
]

2/4 [
Author: Liang Li
Email: liang.li@windriver.com
Subject: WHRV: fsl_8572ds board support
Date: Wed, 28 Apr 2010 13:49:04 +0800

Add support for Freescale 8572ds reference board
running on the Wind River Hypervisor.

The chosen node is needed to enable early console.

Signed-off-by: Liang Li <liang.li@windriver.com>
Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

3/4 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/8572: use common PCI
Date: Mon, 24 May 2010 16:55:28 -0400

Enable the needed pieces to utilize the common PCI
infrastructure that is shared between PPC boards.
The DTS now stores the associated offset which is
used to set up the PCI LAW.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

4/4 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WHRV: fsl_8572ds select PARAVIRT_PTE
Date: Mon, 3 May 2010 11:40:12 -0400

The hypervisor needs the TLB pages to be marked as:
SX/SW/SR/UW/UR/UX

Since the guest OS is running in user mode it needs to
ensure that it sets the pages as (UW/UR/UX).  If this is
not done correctly the hypervisor will interpret these
pages wrong as instructions thus yielding exception 0x400.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>