lttng: fix error offset due to different struct pointer
1/1 [
Author: Zumeng Chen
Email: zumeng.chen@windriver.com
Subject: lttng: fix error offset due to different struct pointer
Date: Mon, 21 Jun 2010 17:45:08 -0700

The result type of macro _va is void pointer, so the offset
is not right, which introduces not proper pgd is processed,
and the expected pgd will never be touched, data exception
will be triggered always at the same place too. So redefine
it with (pgd_t *).

Signed-off-by: Zumeng Chen <zumeng.chen@windriverc.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>