bsp: fsl_p2020ds: fix patch which is being improperly applied
CQID:WIND00220200

When this patch was originally merged the index for the first hunk when
combined with the context which matched in more then one spot resulted
in the lines being added to the wrong part of the code. Here we are reapplying
the patch ensuring the code lands in the right spot thus fixing the index
problem. Generating the patch with addional lines of context will also ensure
this patch does the right thing when applied.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>