powerpc/mm: match variable types to API
1/1 [
Author: Joe MacDonald
Email: joe.macdonald@windriver.com
Subject: powerpc/mm: match variable types to API
Date: Thu, 24 Jun 2010 10:43:21 -0700

sys_subpage_prot() takes an unsigned long for 'addr' then does some stuff
with it and the result is stored in a signed int, i, which is eventually
used as the size parameter in a copy_from_user call.  Update 'i' to be an
unsigned long as well and since 'nm' is used in a size_t context which,
depending on whether this is 32- or 64-bit may be unsigned int or unsigned
long, switch that to a size_t and always be right.

Finally, since we're in the neighbourhood, make the same changes to
subpage_prot_clear().

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>