smc91x: Disable the operation of the SMC91X eeprom
1/1 [
Author: Hongbo Zhong
Email: hongbo.zhong@windriver.com
Subject: smc91x: Disable the operation of the SMC91X eeprom
Date: Wed, 30 Jun 2010 02:30:30 -0700

When you read/write the SMC91X eeprom,the operation will be
fail.at the same time,the network will down on the board of
arm_versatile_926ejs.

  $ ethtool -e eth0

The SMC91X eeprom functionality is not core to the BSP,so
we will simple disable the operation to prevent that the
network device don't work.

Signed-off-by: Hongbo Zhong <hongbo.zhong@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>