rt/cgroups: fix build error
1/1 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: preempt-rt: -rt20 applied to 2.6.34 -standard
Date: Thu, 3 Jun 2010 15:42:03 -0400

This is the complete preempt-rt patchset applied in one single
commit. This corresponds roughly with the -rt20 label referenced
in the upstream project

rt/2.6.33 branch on linux-2.6-tip tree
commit e86825210 [fs: namespace: Fix fuse mount fallout]

with modifications to first apply it to
mainline 2.6.34 (-rt20 is applied to 2.6.33.4) and then move it
on top of our -standard kernel.

To recreate this you must have access to the rt/2.6.34 branch, and
a branch (master in this example) which is Linus' v2.6.34 then
then follow this recipe.

%> git diff master rt/2.6.34 > preempt-rt.patch
%> patch -p1 < preempt-rt.patch

perform merges of any produced *.rej files.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>