oom: protection from the out-of-memory killer
1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: oom: protection from the out-of-memory killer
Date: Mon, 26 Apr 2010 05:56:01 -0700

OOM Protection Feature:

This adds a new capability to the kernel which
may be called (by processes with appropriate priviledges) to
request protection from the out-of-memory killer as long as
the process is using less than the specified amount of virtual
memory.  This is accomplished through a call to setrlimit().

It was possible to get unitialized results returned from
getrlimit/setrlimit calls with OOM_MAXVM flags if passed invalid options
and the limits had not been previously set.  This also corrects a minor
race condition in getrlimit where the structure to be filled out with
OOM_MAXVM data was copied from userspace, it would occasionally cause no
data to be returned.

Child Protection:

When a process is flagged by OOM the first thing that happens is OOM
attempts to kill a child of that process to resolve the out of memory
condition.  It is never considered whether the child process has itself
registered to be protected from OOM.  An example of this happening is when
a number of processes are started manually by a user from a shell.  If one
of the children of shell misbehaves all children will be considered for
termination without seeing whether they are protected.

To address this the same check that is done for a process is added in the
children loop.  For a child if the total number of VM pages allocated are
less than the set value (as set by the child when protecting itself),
the child will not be immediately killed by OOM.

[ This squashes the original two OOM protection patches from the ECGL kernel
  and adds the child protection, all formatted to apply cleanly to the
  2.6.34 kernel as a single patch since they are all related to the same
  overall function. ]

[ The original two patches, as described in "OOM Protection Feature" : ]

Signed-off-by: Yongli he   <yongli.he@windriver.com>
Signed-off-by: MacDonald, Joe <Joe.MacDonald@windriver.com>

[ The child protection function, as described "Child Protection" : ]

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
Signed-off-by: Jason HU <yongqi.hu@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>