preempt_rt: kgdboc: remove invalid list check
1/1 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: preempt_rt: kgdboc: remove invalid list check
Date: Thu, 22 Jul 2010 11:44:20 -0700

During porting of RT to our kernel similar changes were duplicated
to cover the removal of file_move() which included a BUG_ON check for
an empty list. The block of code is correct but the BUG_ON is not needed
since an empty list is valid and expected in this case. Functions which
operate on the list are safe to call with an empty list so no concerns
there.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>