qman: make the bpid of fqid pool to be configurable
1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: qman: make the bpid of fqid pool to be configurable
Date: Mon, 26 Jul 2010 01:54:53 -0700

By default we use bpid 0 as the bman pool for fqid alloc.
But in sAMP when two linux are running at the same time,
there will be a conflict. So we add a new property
"qman-fqalloc-bpid" for qman device node in order to
configure the bpid of fqid pool.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>