lttng: Revert "notifier-atomic-call-chain-notrace"
1/2 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: lttng: Revert "notifier-atomic-call-chain-notrace"
Date: Mon, 26 Jul 2010 01:11:18 -0700

This reverts commit notifier-atomic-call-chain-notrace.

The original idea of the commit is marking the rcu_read_lock/unlock() as
notrace, but it introduced the preemption disabling for it called
rcu_read_lock/unlock_sched_notrace(), but disabling preemption is not necessary
and will introduce latency and also potential "BUG: sleeping function called
from invalid context".

To mark the rcu_read_lock/unlock() as notrace directly is not easier, for it
has lots of callees, we can not simply mark all of them as notrace, but as we
know, notrace means stop the tracing, so, we get the key to this problem: just
call tracing_stop() and tracing_start() to do the job instead of marking them
with notrace. A new patch will include the solution.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

2/2 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: lttng: stop tracing of the notifier atomic call chain
Date: Mon, 26 Jul 2010 01:11:19 -0700

The last patch reverts the commit 3a6c5a5b6f13e6e4c29bc2165b02cea4d6e8141c,
This patch re-implement the idea of the commit but avoid introducing preemption
disabling via calling tracing_stop/start() instead of
rcu_read_lock/unlock_sched_notrace() and also we will not try to implement the
notrace variant of rcu_read_lock/unlock() for it's hard to convert all of the
callees of them to notrace.

With the last patch and this one, the un-necessary preemption disabling block
is removed and the pontential "BUG: sleeping function called from invalid
context" is also fixed.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>