WRHV: Set decrementer based on new HV mdio
1/1 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV: Set decrementer based on new HV mdio
Date: Thu, 5 Aug 2010 21:44:54 -0400

Previously the decrementer was set incorrectly and boards
which relied on it for things like the sleep command were
not getting expected results.

The hypervisor was not providing the CPU freq. correctly to
the guest and /proc/cpu was not correct compared to that of
native.  What was being used for CPU freq. was actually the
board clock freq.

Double check that the XML settings for TimerTickFrequency is the
same as CONFIG_HZ by comparing it to VBI_TICK_TIMER_FREQ_GET, which
is the tick freq provide in the HV config structure.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>