ppc-mc/e500: fix incorrect debug exception entry
1/1 [
Author: Tiejun Chen
Email: tiejun.chen@windriver.com
Subject: ppc-mc/e500: fix incorrect debug exception entry
Date: Fri, 13 Aug 2010 17:29:56 +0800

When support SPE we introduce a dummy handler in linux kernel in order
to create the correct SPE interrupt vector offset as the hypervisor. But
actually that should be used for debug exception is defined by guest OS.

Now fix that to be dedicated to DEBUG exception.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>