mips: disable subsection optimization in atomic_add when profiling
1/1 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: mips: disable subsection optimization in atomic_add when profiling
Date: Sun, 15 Aug 2010 14:00:14 -0400

The current implementation of atomic_add uses .subsection
directives to move the unlikely condition out of line, in
a similar way that spinlocks do.  But there is an assumption
that the out of line code won't be moved more than 128k away.

Normally we are building with --ffunction-sections, so this isn't
a concern.  But gcc will forcibly disable this option with profiling
enabled, even if you hard code it in the Makefile, e.g:

linux/fs/ocfs2/alloc.c:1: warning: -ffunction-sections disabled; it makes profiling impossible

and on big files like alloc.c with sizes like this:

ocfs2$size alloc.o
   text	   data	    bss	    dec	    hex	filename
 137304	    400	      0	 137704	  219e8	alloc.o
ocfs2$

you break the 128k limit and get these errors

{standard input}: Assembler messages:
{standard input}:871: Error: Branch out of range

which trace back to atomic_add().  The use of subsection in
atomic_add() isn't necessarily a significant optimization, see:

	http://www.spinics.net/lists/mips/msg34820.html

and so disabling it when CONFIG_PROFILING is enabled allows things to
build without failing.  There isn't a hard 1:1 mapping between
CONFIG_PROFILING and the use of "gcc -p ..." but CONFIG_PROFILING is
largely unused, and the use of it here for this seeems to make sense.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>