audit: fix kernel panic due to invalid syscall
1/1 [
Author: Lans Zhang
Email: jia.zhang@windriver.com
Subject: audit: fix kernel panic due to invalid syscall
Date: Mon, 16 Aug 2010 21:04:19 -0700

A program making a syscall with an invalid syscall number can cause kernel
panic. The audit code uses a hash table with syscall number used to index
hash table, but it only handles up to 2048 entries. Therefore, with a large
syscall number, an invalid bucket in the hash table is accessed, and the kernel
panics or sometimes just hangs(which is worse). To fix the problem, just add
the range check for syscall number.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>