preempt_rt: Fix the usage of init_MUTEX*
1/1 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: preempt_rt: Fix the usage of init_MUTEX*
Date: Fri, 13 Aug 2010 06:24:21 -0700

init_MUTEX* should be converted to sema_init() in preempt_rt, this patch
fixes all of the related usage.

Without this patch, we may get the error of "implicit declaration of
function 'init_MUTEX'".

And a wrong usage of sema_init() is fixed in this patch too.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>