fix kernel panic due to invalid syscall
1/1 [
Author: Lans Zhang
Email: jia.zhang@windriver.com
Subject: fix kernel panic due to invalid syscall
Date: Wed, 18 Aug 2010 20:21:08 -0700

A program making a syscall with an invalid syscall number can cause kernel
panic. The audit code uses a hash table with syscall number used to index
hash table, but it only handles up to 2048 entries. Therefore, with a large
syscall number, an invalid bucket in the hash table is accessed, and the kernel
panics or sometimes just hangs(which is worse). To fix the problem, just add
the range check for syscall number. Additionally, we need a NR_syscalls macro
for arm and mips under cgl configuration, because the check is placed at
common area and other architectures all have NR_syscalls.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>