preempt_irq: USB fixup is not CGL specific.
The USB patch is not specific to CGL, but preempt_irq is CGL
specific. Move the patch to the "irq" feature, so that builds
of standard+feature/irq will also get the patch.

Note that the ordering of patching for CGL during a treegen
should remain unchanged, since it moved from the head of
preempt_irq to the tail of irq, and preempt_irq starts by
including irq.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>