wrhv/powerpc: enable SPE support for E500 core
1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/powerpc: enable SPE support for E500 core
Date: Mon, 31 May 2010 16:28:08 +0800

Add SPE support for E500 core. Currently in hypervisor the
debug exception is defined by default. So we have to add a
dummy handler in linux kernel in order to create the correct
interrupt vector offset as the hypervisor.

When support SPE we introduce a dummy handler in linux kernel in order
to create the correct SPE interrupt vector offset as the hypervisor. But
actually that should be used for debug exception is defined by guest OS.

Now fix that to be dedicated to DEBUG exception.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>