preempt_rt: x86: trace_hardirqs_off/on avoid IMV optimazation
1/1 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: preempt_rt: x86: trace_hardirqs_off/on avoid IMV optimazation
Date: Fri, 20 Aug 2010 12:51:06 -0700

Use the variant of trace_preemptirqsoff_hist() which does not make use
of the immediate value optimizations (ie. the one that starts
with an underscore) for the preemptirqsoff tracer. If we do use the IMV
optimization, attempting the text_poke which places the
BREAKPOINT_INSTRUCTION in arch_imv_update() will cause a fault which
will reboot the system.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>