wrhv/dpa: fix the dpa_bus_to_virt bug when PAMU is enabled
1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/dpa: fix the dpa_bus_to_virt bug when PAMU is enabled
Date: Tue, 17 Aug 2010 01:27:31 -0700

When PAMU is disabled, the physical address is equal to dma
address in guest OS. But when PAMU is enabled, maybe the
physical address and dma address are different. And we have
to use the dma address of PAGE_OFFSET to get the correct
virtual address.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>