sctp: fix oops with unconnected sockets
1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: sctp: fix oops with unconnected sockets
Date: Tue, 24 Aug 2010 17:08:39 -0400

In a few boundary cases a socket could be used where it isn't in a
connected or listening state for SCTP traffic.  When this occurs
the attached network is NULL which in-turn causes the __xfrm_lookup()
function to panic by dereferencing the NULL net.

To address this a check is added in the SCTP get_dst() function where
if a net attached to a socket is NULL, the function simply returns
NULL indicating no destination for the intended SCTP association as
the socket itself has no connection.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>