preempt_rt: x86: trace_hardirqs_off/on avoid IMV optimization
1/1 [
Author: Matt Sherer
Email: matt.sherer@windriver.com
Subject: preempt_rt: x86: trace_hardirqs_off/on avoid IMV optimization
Date: Thu, 2 Sep 2010 14:56:07 -0700

Use the variant of trace_preemptirqsoff_hist() which does not make use
of the immediate value optimizations (ie. the one that starts
with an underscore) for the preemptirqsoff tracer. If we do use the IMV
optimization, attempting the text_poke which places the
BREAKPOINT_INSTRUCTION in arch_imv_update() will cause a fault which
will reboot the system.

This patch was in place previously for other tracing calls but needs
to be extended to cover trace_hardirqs_o{n,ff}_caller().
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>