tracing/ftrace: add more arch specific support around modifying code
1/2 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: tracing/ftrace: add more arch specific support around modifying code
Date: Thu, 2 Sep 2010 03:52:27 -0700

ftrace_arch_code_modify_prepare/post_process() are added to do arch
specific operations before/after modifying code in kernel, Similarly,
this patch adds ftrace_arch_module_modify_prepare/post_process()
before/after modifying the code in module.

To distinguish the different opertions for different commands, the
command argument is passed to ftrace_arch_code_modify_post_process() and
accordingly, the command flags are moved to ftrace.h for the archs.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

2/2 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: tracing/ftrace: MIPS: Fixup of the support for SMP
Date: Thu, 2 Sep 2010 03:52:28 -0700

To flush the icaches on the other cpus, the irqs of the cpus should be
enabled to accept the ipi interrupts. otherwise, the icaches on the
other cpus may not be flushed and even make the system hang: the current
cpu may wait for the other cpus to flush their icaches, but the other
cpus can not response for the irqs are disabled.

To fix this problem, this patch delays the icache flushing operation
after the irqs are enabled.

Benefit from this fix, the system performance will be improved for
currently, we only need to flush the whole icache one time at the end
instead of flush thousands of entries individually.

A 'side effect' is, from now on, any user of ftrace_modify_code() in
MIPS should flush the icache itself.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>