kexec: fix memory calculation problem
1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: kexec: fix memory calculation problem
Date: Tue, 7 Sep 2010 13:29:53 -0700

For kexec early in the boot we check to ensure that there is enough
free memory to proceed.  Previously this was done by calculating against
the value MAX_MEMORY.  A recent fix corrected the 'mem=' parameter but
also changed the definition of MAX_MEMORY to now be the amount of DRAM
available on the system.  This broke the kexec calculation preventing
any memory from being free'd during a kexec fast reboot.

To address this the calculation is being changed to simply check that
there is at least 256M free memory available, which has been deemed
to be sufficient.  If less than that is available, the code will
cleanup any memory from the previous kernel before proceeding.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>