RMI/kexec: add kexec support for RPM XLS-based boards
1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: RMI/kexec: add kexec support for RPM XLS-based boards
Date: Fri, 10 Sep 2010 09:32:15 -0700

Main feature needed is the specific SMP support. Secondary
cores/threads spin on tight loops in the bootloader memory:
kexec support has to mimic this when doing the handoff from
the first kernel to the second. Unfortunately, this has the
side-effect of preventing from booting a second kernel with
more CPUs than the first one. The opposite works though, as
a crash dump kernel should really be booted with only one CPU
to use an amount of memory as small as possible.

Also included is code to reset devices upon booting.  For kexec to
work properly, the message rings for each CPU must be shut off,
the pending interrupts cleared and ACK'ed, the message ring timers
shut-off and the GMACs reset.

These changes are derived from the kexec support for the XLR732-based
boards completed in the previous release.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>