Squash development history from preempt_rt
There were about 10 commits to the preempt_rt ktype which were
small changes to get the original large preempt_rt patch to work
correctly. This essentially equates to development history which
is of no use to consumers of our kernel, so we are squashing these
into the main preempt-rt patch. In addition we are squashing the
VFS revert patch which was merged into -tip after -rt20, this
vastly simplifies the large preempt_rt patch and removes the need
for several of the 10 patches mentioned above.

You might observe that there are 10 patch removals but you will
see 11 from preempt_rt.scc, this is because one patch was
incorrectly listed twice prior to this commit.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>