ftrace: integrate patch 1/8 from Wu into existing patch
CQID: WIND00232718

Patch 1 of Wu's series needs to be folded into an existing
patch, since it was a fix of a prev. introduced error.

The stand-alone 1/8 was:

---------------
With the wrong condition check in
ftrace_arch_code_modify_post_process(), some of the kernel entries will
not be flushed and make the selftest of dynamic ftracer fails.

The *command may be the OR result of the following values:

include/linux/ftrace.h:

	FTRACE_ENABLE_CALLS		= (1 << 0),
	FTRACE_DISABLE_CALLS		= (1 << 1),
	FTRACE_UPDATE_TRACE_FUNC	= (1 << 2),
	FTRACE_ENABLE_MCOUNT		= (1 << 3),
	FTRACE_DISABLE_MCOUNT		= (1 << 4),
	FTRACE_START_FUNC_RET		= (1 << 5),
	FTRACE_STOP_FUNC_RET		= (1 << 6),

If the FTRACE_ENABLE_CALLS or FTRACE_DISABLE_CALLS, we must flush the
whole icache for we don't know which kernel entries are modified. if
only one of the others is set, we just need to flush the related entry.

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
---
 arch/mips/kernel/ftrace.c |   13 +++++++------
 1 files changed, 7 insertions(+), 6 deletions(-)

diff --git a/arch/mips/kernel/ftrace.c b/arch/mips/kernel/ftrace.c
index 5d21901..785965b 100644
--- a/arch/mips/kernel/ftrace.c
+++ b/arch/mips/kernel/ftrace.c
@@ -85,15 +85,16 @@ int ftrace_arch_code_modify_post_process(void *data)
 {
 	int *command = data;

+	if (*command & (FTRACE_ENABLE_CALLS | FTRACE_DISABLE_CALLS)) {
+		__flush_cache_all();
+		return 0;
+	}
+
 	if (*command & FTRACE_UPDATE_TRACE_FUNC)
 		flush_icache_from(FTRACE_CALL_IP);
-#ifdef CONFIG_FUNCTION_GRAPH_TRACER
-	else if ((*command & FTRACE_START_FUNC_RET)
-		|| (*command & FTRACE_STOP_FUNC_RET))
+
+	if (*command & (FTRACE_START_FUNC_RET | FTRACE_STOP_FUNC_RET))
 		flush_icache_from(FTRACE_GRAPH_CALL_IP);
-#endif
-	else
-		__flush_cache_all();

 	return 0;
 }
--------------------

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>